fix(gun): FID-0088 shell-casing tint uses well-defined u32 shifts
The faithful shell-casing envcolour path shifted u8 tileColor fields
left by 24 without the (u32) cast its sibling legacy path already has,
which is signed-overflow UB for byte values >= 0x80 (identical bits on
every real compiler, but inconsistent with the sibling). Add the casts
so both branches are well-defined. No behavior change; bit-identical.

+++ src/game/gun.c
@@ -31051,10 +31051,10 @@ void sub_GAME_7F068EC4(CasingRecord *entry, Gfx **gdl) {
                 ((u32)cp[0xFDC] << 24) | ((u32)cp[0xFDD] << 16) |
                 ((u32)cp[0xFDE] << 8) | (u32)cp[0xFDF];
         } else {
-            renderdata.envcolour.word = (g_CurrentPlayer->tileColor.r << 24) |
-                                        (g_CurrentPlayer->tileColor.g << 16) |
-                                        (g_CurrentPlayer->tileColor.b << 8) |
-                                        g_CurrentPlayer->tileColor.a;
+            renderdata.envcolour.word = ((u32)g_CurrentPlayer->tileColor.r << 24) |
+                                        ((u32)g_CurrentPlayer->tileColor.g << 16) |
+                                        ((u32)g_CurrentPlayer->tileColor.b << 8) |
+                                        (u32)g_CurrentPlayer->tileColor.a;
         }
 
         subdraw(&renderdata, &model);
